Update batch_bagger.py
Unpack all bags in the target directory, and delete files and directories created in the bagging process. Prompt user to confirm this prior to running, as this process will invalidate all bags. Fix syntax for reading CSV and XLSX files for a list of bags to unpack. Test to avoid attempting to unpack non-bag directories when no spreadsheet list of bags is provided, to prevent these directories from throwing an error.